fixes for schema merging flow fixes for config parsing fixes for error management added more cli tests
